factor pull target into pull_targets, change interface of source/target parsing from cmd to config
