+++ pull_target.go
@@ -1,0 +1,156 @@
+package main
+
+import (
+	"fmt"
+	"strings"
+
+	"gopkg.in/yaml.v2"
+
+	"github.com/phrase/phraseapp-go/phraseapp"
+)
+
+type Targets []*Target
+
+func (targets Targets) ProjectIds() []string {
+	projectIds := []string{}
+	for _, target := range targets {
+		projectIds = append(projectIds, target.ProjectID)
+	}
+	return projectIds
+}
+
+type Target struct {
+	File          string
+	ProjectID     string
+	AccessToken   string
+	FileFormat    string
+	Params        *PullParams
+	RemoteLocales []*phraseapp.Locale
+}
+
+func (tgt *Target) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	m := map[string]interface{}{}
+	err := phraseapp.ParseYAMLToMap(unmarshal, map[string]interface{}{
+		"file":         &tgt.File,
+		"project_id":   &tgt.ProjectID,
+		"access_token": &tgt.AccessToken,
+		"file_format":  &tgt.FileFormat,
+		"params":       &m,
+	})
+	if err != nil {
+		return err
+	}
+
+	tgt.Params = new(PullParams)
+	if v, found := m["locale_id"]; found {
+		if tgt.Params.LocaleID, err = phraseapp.ValidateIsString("params.locale_id", v); err != nil {
+			return err
+		}
+		// Must delete the param from the map as the LocaleDownloadParams type
+		// doesn't support this one and the apply method would return an error.
+		delete(m, "locale_id")
+	}
+	return tgt.Params.ApplyValuesFromMap(m)
+
+}
+
+func (target *Target) CheckPreconditions() error {
+	if err := ValidPath(target.File, target.FileFormat, ""); err != nil {
+		return err
+	}
+
+	if strings.Count(target.File, "*") > 0 {
+		return fmt.Errorf(
+			"File pattern for 'pull' cannot include any 'stars' *. Please specify direct and valid paths with file name!\n %s#targets", docsConfigUrl,
+		)
+	}
+
+	duplicatedPlaceholders := []string{}
+	for _, name := range []string{"<locale_name>", "<locale_code>", "<tag>"} {
+		if strings.Count(target.File, name) > 1 {
+			duplicatedPlaceholders = append(duplicatedPlaceholders, name)
+		}
+	}
+
+	if len(duplicatedPlaceholders) > 0 {
+		dups := strings.Join(duplicatedPlaceholders, ", ")
+		return fmt.Errorf(fmt.Sprintf("%s can only occur once in a file pattern!", dups))
+	}
+
+	if target.GetLocaleID() == "" && !containsAnyPlaceholders(target.File) {
+		return fmt.Errorf("Could not find any locale information. Please specify a 'locale_id' in your params or provide a placeholder!")
+	}
+
+	return nil
+}
+
+func (target *Target) localeForRemote() (*phraseapp.Locale, error) {
+	for _, locale := range target.RemoteLocales {
+		if locale.ID == target.GetLocaleID() || locale.Name == target.GetLocaleID() {
+			return locale, nil
+		}
+	}
+	return nil, fmt.Errorf("Provided locale_id %q but did not match any remote locales in project %q", target.GetLocaleID(), target.ProjectID)
+}
+
+func (t *Target) GetFormat() string {
+	if t.Params != nil && t.Params.FileFormat != nil {
+		return *t.Params.FileFormat
+	}
+	if t.FileFormat != "" {
+		return t.FileFormat
+	}
+	return ""
+}
+
+func (t *Target) GetLocaleID() string {
+	if t.Params != nil {
+		return t.Params.LocaleID
+	}
+	return ""
+}
+
+func (t *Target) GetTag() string {
+	if t.Params != nil && t.Params.Tag != nil {
+		return *t.Params.Tag
+	}
+	return ""
+}
+
+func TargetsFromConfig(config phraseapp.Config) (Targets, error) {
+	if config.Targets == nil || len(config.Targets) == 0 {
+		return nil, fmt.Errorf("no targets for download specified")
+	}
+
+	tmp := struct {
+		Targets Targets
+	}{}
+	err := yaml.Unmarshal(config.Targets, &tmp)
+	if err != nil {
+		return nil, err
+	}
+	tgts := tmp.Targets
+
+	projectId := config.DefaultProjectID
+	fileFormat := config.DefaultFileFormat
+
+	validTargets := []*Target{}
+	for _, target := range tgts {
+		if target == nil {
+			continue
+		}
+		if target.ProjectID == "" {
+			target.ProjectID = projectId
+		}
+		if target.FileFormat == "" {
+			target.FileFormat = fileFormat
+		}
+		validTargets = append(validTargets, target)
+	}
+
+	if len(validTargets) <= 0 {
+		return nil, fmt.Errorf("no targets could be identified! Refine the targets list in your config")
+	}
+
+	return validTargets, nil
+}
